Lesson-1 Develop a basic hapijs application.
This basic hapijs application has two plugins each with one route.
These two plugins illustrate hapijs' modular organization.  There are two plugins in
the application. Plugin "one" contains the web servers' index route.
And, plugin "two" contains a route to a sub domain named "boom".
There is one path in the subdomain which is index.

Additionally, this application builds two nodejs servers.
One, is to be web server and is called "web".  The other is to
be a web service api and is called "api". This shows the unique power
hapijs / nodejs offers the developer. The api is built here but no
code will be written for it until later.  For now we are going to
focus on the web application development of hapijs.

This is the cleanest modular structure I have ever seen for
creating web applications and web apis. After you develop this
basic hapijs server I believe you will agree.